+++ apps/common/drf/parsers/__init__.py
@@ -1,1 +1,2 @@
 from .csv import *
+from .excel import *
